Edit Post: Expose PluginSidebar as the final API
