+++ edit-post/components/sidebar/index.js
@@ -1,7 +1,7 @@
 /**
  * WordPress Dependencies
  */
-import { ifCondition, withFocusReturn } from '@wordpress/components';
+import { createSlotFill, ifCondition, withFocusReturn } from '@wordpress/components';
 import { withSelect } from '@wordpress/data';
 import { compose } from '@wordpress/element';
 
@@ -10,6 +10,8 @@ import { compose } from '@wordpress/element';
  */
 import './style.scss';
 
+const SidebarFill = createSlotFill( 'editPost.Sidebar' );
+
 /**
  * Renders a sidebar with its content.
  *
@@ -17,21 +19,27 @@ import './style.scss';
  */
 const Sidebar = ( { children, label } ) => {
 	return (
-		<div
-			className="edit-post-sidebar"
-			role="region"
-			aria-label={ label }
-			tabIndex="-1"
-		>
-			{ children }
-		</div>
+		<SidebarFill>
+			<div
+				className="edit-post-sidebar"
+				role="region"
+				aria-label={ label }
+				tabIndex="-1"
+			>
+				{ children }
+			</div>
+		</SidebarFill>
 	);
 };
 
-export default compose(
+const WrappedSidebar = compose(
 	withSelect( ( select ) => ( {
 		activeSidebarName: select( 'core/edit-post' ).getActiveGeneralSidebarName(),
 	} ) ),
 	ifCondition( ( { activeSidebarName, name } ) => activeSidebarName === name ),
 	withFocusReturn,
 )( Sidebar );
+
+WrappedSidebar.Slot = SidebarFill.Slot;
+
+export default WrappedSidebar;
